requirements.txt
numpy==2.2.0
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
pytesseract==0.3.13
